Implement a simple count based boot marker
We write the number of failed boots to the marker and only
prune the dalvik cache if the number of consecutive failed
boots is higher than a predefined threshold.

Note that the code is forgiving of errors related to boot
markers; we continue the boot process even if we're unable
to create / write or flush the boot marker.

CRs-Fixed: 871040

Change-Id: I51d2961bc93f00ce6f4e55de184b54db1db00f23